+++ tests/serial/explicit/TestExplicitMPI.cpp
@@ -1,0 +1,21 @@
+#ifndef PRECICE_NO_MPI
+
+#include <precice/SolverInterface.hpp>
+#include <vector>
+#include "helpers.hpp"
+#include "testing/Testing.hpp"
+
+BOOST_AUTO_TEST_SUITE(PreciceTests)
+BOOST_AUTO_TEST_SUITE(Serial)
+BOOST_AUTO_TEST_SUITE(Explicit)
+BOOST_AUTO_TEST_CASE(TestExplicitMPI)
+{
+  PRECICE_TEST("SolverOne"_on(1_rank), "SolverTwo"_on(1_rank));
+  runTestExplicit(context.config(), context);
+}
+
+BOOST_AUTO_TEST_SUITE_END() // PreciceTests
+BOOST_AUTO_TEST_SUITE_END() // Serial
+BOOST_AUTO_TEST_SUITE_END() // Explicit
+
+#endif // PRECICE_NO_MPI
