+++ tests/serial/explicit/helpers.hpp
@@ -1,0 +1,12 @@
+#ifndef PRECICE_NO_MPI
+
+#pragma once
+
+#include "testing/TestContext.hpp"
+
+using namespace precice;
+using precice::testing::TestContext;
+
+void runTestExplicit(std::string const &configurationFileName, TestContext const &context);
+
+#endif
